requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
snowflake-sqlalchemy
python-dotenv
requests
dbt-core
dbt-snowflake
pydantic
